Now allowing maximum functors to be 2**16 - 1.	Various	changes to make this work properly for pure SLG execution.  Much of this simply involved changing the occurrences of byte types in psc-oriented routines to int.  While I was at it, I changed the name from the confusing insert() to insert_psc().
No attempt was made to expand bytecode,	so large arity terms are
not available for explicit compilation.	 Also, I dodn't	fiddle
around with assert/retract.  However, nearly all Prolog	builtins
have been checked to ensure they work with large-arity terms, as
have been interned tries.  I started working on	checking tabling,
although there are more	features to check.

For good luck, I also increased	the HEAP_GC_MARGIN.

Fixed is_most_general_term to work with	large terms (lists as well as functors)




git-svn-id: svn+ssh://svn.code.sf.net/p/xsb/src/trunk@8944 756869e0-9e47-4bc5-b0a8-847305dc5e1c

+++ XSB/emu/error_xsb.c
@@ -203,7 +203,7 @@ DllExport void call_conv xsb_throw_internal(CTXTdeclc prolog_term Ball, size_t B
   }
   */
 
-  exceptballpsc = pair_psc((Pair)insert("$$exception_ball", (byte)2, 
+  exceptballpsc = pair_psc((Pair)insert("$$exception_ball", 2, 
 					pair_psc(insert_module(0,"standard")), 
 					&isnew));
   hreg_start = hreg;
@@ -303,7 +303,7 @@ DllExport void call_conv xsb_throw(CTXTdeclc prolog_term Ball)
 	    ctrace_ctr++);
   }
 
-  exceptballpsc = pair_psc((Pair)insert("$$exception_ball", (byte)2, 
+  exceptballpsc = pair_psc((Pair)insert("$$exception_ball", 2, 
 					pair_psc(insert_module(0,"standard")), 
 					&isnew));
   hreg_start = hreg;
